refactor(data): 상품 목 서버 분리
